refactor: extract tls config from urlfetchers to ensure the same config is use for all the workflow
